changed int i to unsigned long
Warning error of int being compared to unsigned long. Errors are removed with this fix. 
Note: Trying running on windows to check fix.

+++ HW2/main.cpp
@@ -49,7 +49,7 @@ int main()
  int parentheses_are_valid(string input)
 {
     int parentheses = 0;
-    for ( int i = 0; i < input.size() ; i ++)
+    for ( unsigned long i = 0; i < input.size() ; i ++)
     {
         if (input.at(i) == '(')
         {
